+++ safecopy/__init__.py
@@ -1,5 +1,7 @@
 # SafeCopy package
 
-"""SafeCopy - Automated Backup Tool"""
+"""
+SafeCopy - Automated Backup Tool
+"""
 
-__version__ = "0.1.0"
+__version__ = "0.2.0"